backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.1
defusedxml==0.8.0rc2
Django==3.2.16
django-autoslug==1.9.9
django-colorfield==0.11.0
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
django-cors-headers==3.13.0
djoser==2.1.0
fonttools==4.49.0
fpdf2==2.7.8
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.4
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
psycopg2-binary==2.9.3
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.8
